+++ e2e/fixtures/fixtures.js
@@ -1,7 +1,23 @@
 export const isType = "module.exports = function isType() { return 'got is-type'; };";
+export const isTypeSpec = testShouldPass => `const expect = require('chai').expect;
+const isType = require('./is-type.js');
+
+describe('isType', () => {
+  it('should display "got is-type"', () => {
+    expect(isType())${testShouldPass ? '' : '.not'}.to.equal('got is-type');
+  });
+});`;
 export const isTypeES6 = "export default function isType() { return 'got is-type'; };";
 export const isString =
   "const isType = require('./is-type.js'); module.exports = function isString() { return isType() +  ' and got is-string'; };";
+export const isStringSpec = testShouldPass => `const expect = require('chai').expect;
+const isString = require('./is-string.js');
+
+describe('isString', () => {
+  it('should display "got is-type and got is-string"', () => {
+    expect(isString())${testShouldPass ? '' : '.not'}.to.equal('got is-type and got is-string');
+  });
+});`;
 export const isStringES6 =
   "import isType from './is-type.js'; export default function isString() { return isType() +  ' and got is-string'; };";
 export const barFooFixture =