+++ src/consumer/consumer.js
@@ -522,15 +522,9 @@ export default class Consumer {
     );
   }
 
-  async listComponentsForAutoTagging(modifiedComponents: BitId[]) {
+  async listComponentsForAutoTagging(modifiedComponents: BitId[]): Promise<ModelComponent[]> {
     const candidateComponents = await this.candidateComponentsForAutoTagging(modifiedComponents);
     return this.scope.bumpDependenciesVersions(candidateComponents, modifiedComponents, false);
-  }
-
-  async bumpDependenciesVersions(committedComponents: Component[]) {
-    const committedComponentsIds = committedComponents.map(committedComponent => committedComponent.id);
-    const candidateComponents = await this.candidateComponentsForAutoTagging(committedComponentsIds);
-    return this.scope.bumpDependenciesVersions(candidateComponents, committedComponentsIds, true);
   }
 
   /**
@@ -688,7 +682,7 @@ export default class Consumer {
       });
       if (!R.isEmpty(componentsWithMissingDeps)) throw new MissingDependencies(componentsWithMissingDeps);
     }
-    const committedComponents = await this.scope.putMany({
+    return this.scope.putMany({
       consumerComponents: components,
       message,
       exactVersion,
@@ -697,9 +691,6 @@ export default class Consumer {
       consumer: this,
       verbose
     });
-    const autoUpdatedComponents = await this.bumpDependenciesVersions(committedComponents);
-
-    return { components, autoUpdatedComponents };
   }
 
   static getNodeModulesPathOfComponent(bindingPrefix: string, id: BitId): PathOsBased {